+++ src/app/notifications/page.tsx
@@ -163,7 +163,7 @@ export default function NotificationsPage() {
               {t("notifications.noChannels")}
             </p>
             <Button asChild variant="outline">
-              <Link href="/settings">
+              <Link href="/settings/notifications">
                 <Settings className="mr-2 h-4 w-4" />
                 {t("notifications.goToSettings")}
               </Link>
