+++ src/app/settings/[section]/page.tsx
@@ -1,0 +1,45 @@
+import { notFound } from "next/navigation";
+
+import { SectionPlaceholder } from "@/components/settings/section-placeholder";
+import {
+  SETTINGS_SECTION_SLUGS,
+  isSettingsSectionSlug,
+} from "@/components/settings/section-slugs";
+import { SettingsShell } from "@/components/settings/settings-shell";
+
+/**
+ * Dynamic settings section route. Each of the eight `SETTINGS_SECTION_SLUGS`
+ * is pre-rendered at build via `generateStaticParams()` so the URLs are
+ * statically known to Next.js, while the `dynamicParams = false` flag below
+ * tells the router to 404 (instead of attempting on-demand rendering) for any
+ * slug not in the list — which is exactly what `notFound()` would do at
+ * request time, just earlier and without rendering.
+ */
+
+export const dynamicParams = false;
+
+export function generateStaticParams() {
+  return SETTINGS_SECTION_SLUGS.map((section) => ({ section }));
+}
+
+interface PageProps {
+  // Next.js 16 made route `params` an async Promise. We `await` it before use.
+  params: Promise<{ section: string }>;
+}
+
+export default async function SettingsSectionPage({ params }: PageProps) {
+  const { section } = await params;
+
+  // Defence-in-depth — `dynamicParams = false` already 404s unknown slugs at
+  // routing time, but we re-check here so a hand-rolled override of the route
+  // config can never silently fall through to a typo'd slug.
+  if (!isSettingsSectionSlug(section)) {
+    notFound();
+  }
+
+  return (
+    <SettingsShell active={section}>
+      <SectionPlaceholder slug={section} />
+    </SettingsShell>
+  );
+}
