+++ src/components/settings/section-slugs.ts
@@ -1,0 +1,26 @@
+/**
+ * Server-safe constants for the v1.4 settings split. Kept in a separate
+ * module from `<SettingsShell>` (a client component) so server-rendered code
+ * paths — `generateStaticParams()`, the page guard, future tests — can import
+ * the slug list without dragging the whole client bundle into the server
+ * graph.
+ */
+
+export const SETTINGS_SECTION_SLUGS = [
+  "account",
+  "integrations",
+  "notifications",
+  "dashboard",
+  "ai",
+  "api",
+  "advanced",
+  "about",
+] as const;
+
+export type SettingsSectionSlug = (typeof SETTINGS_SECTION_SLUGS)[number];
+
+export function isSettingsSectionSlug(
+  value: string,
+): value is SettingsSectionSlug {
+  return (SETTINGS_SECTION_SLUGS as readonly string[]).includes(value);
+}
